Clean up and isolate remaining code
